+++ kedro/ipython/logo-32x32.png
